recognition-service/pyproject.toml
[tool.poetry]
name = "recognition-service"
version = "0.1.0"
description = ""
authors = ["Matt Kirby <MattJKirby@outlook.com>"]
readme = "README.md"
packages = [{include = "recognition_service"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
flask = "^2.2.2"
flask-sqlalchemy = "^3.0.2"
pillow = "^9.2.0"
numpy = "^1.23.4"
matplotlib = "^3.6.1"
pymongo = "^4.3.2"
tensorflow-macos = "^2.10.0"
tensorflow-metal = "^0.6.0"
scikit-learn = "^1.1.3"
flask-cors = "^3.0.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
